requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-filter==23.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
